requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
orjson
loguru
tenacity
SQLAlchemy
alembic
psycopg[binary]
openai
langchain
langchain-openai
langgraph
tiktoken
python-multipart
python-dotenv